Added files 'Responses' and 'Server'

+++ config/setup.js
@@ -1,6 +1,14 @@
 import 'dotenv/config'
 import supertest from "supertest"
 import ConfigHelper from "../helpers/config.helper";
+import {start} from './server'
+
+const baseUrl = process.env.BASE_URL
+const port = process.env.PORT
+const isMock = baseUrl.includes('localhost') && baseUrl.includes(port)
+
+if(isMock)
+    start(port)
 
 before(async function () {
    let response = await supertest(process.env.BASE_URL)
@@ -10,6 +18,8 @@ before(async function () {
 })
 
 after(async function(){
-    const configHelper = new ConfigHelper()
-    await configHelper.delete()
+    if(!isMock) {
+        const configHelper = new ConfigHelper()
+        await configHelper.delete()
+    }
 })
